+++ src/app/models/Product.ts
@@ -1,0 +1,59 @@
+import mongoose, { Schema, Document, models } from 'mongoose';
+
+export interface IProduct extends Document {
+  title: string;
+  description: string;
+  price: number;
+  category: string;
+  image: string;
+  inStock: boolean;
+  isDeleted: boolean;
+  deletedAt: Date | null;
+  createdAt: Date;
+  updatedAt: Date;
+}
+
+const ProductSchema: Schema<IProduct> = new Schema(
+  {
+    title: {
+      type: String,
+      required: [true, 'Product title is required'],
+      trim: true,
+    },
+    description: {
+      type: String,
+      required: [true, 'Product description is required'],
+    },
+    price: {
+      type: Number,
+      required: [true, 'Product price is required'],
+      min: [0, 'Price must be positive'],
+    },
+    category: {
+      type: String,
+      required: [true, 'Product category is required'],
+    },
+    image: {
+      type: String,
+      required: [true, 'Product image is required'],
+    },
+    inStock: {
+      type: Boolean,
+      default: true,
+    },
+    isDeleted: {
+      type: Boolean,
+      default: false,
+    },
+    deletedAt: {
+      type: Date,
+      default: null,
+    },
+  },
+  {
+    timestamps: true,
+  }
+);
+
+export const Product =
+  models.Product || mongoose.model<IProduct>('Product', ProductSchema);
