+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Image from "next/image";
+
 
 export default function Home() {
   return (
